requirements.txt
pygpgme>=0.3
nose>=1.1.2
mock>=1.0b1
